client_part
