--- openstack/obs/client.go
@@ -1,190 +1,0 @@
-package obs
-
-import (
-	"errors"
-	"io"
-	"os"
-	"sort"
-	"strings"
-)
-
-func (obsClient ObsClient) ListMultipartUploads(input *ListMultipartUploadsInput) (output *ListMultipartUploadsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ListMultipartUploadsInput is nil")
-	}
-	output = &ListMultipartUploadsOutput{}
-	err = obsClient.doActionWithBucket("ListMultipartUploads", HTTP_GET, input.Bucket, input, output)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
-func (obsClient ObsClient) AbortMultipartUpload(input *AbortMultipartUploadInput) (output *BaseModel, err error) {
-	if input == nil {
-		return nil, errors.New("AbortMultipartUploadInput is nil")
-	}
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-	output = &BaseModel{}
-	err = obsClient.doActionWithBucketAndKey("AbortMultipartUpload", HTTP_DELETE, input.Bucket, input.Key, input, output)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
-func (obsClient ObsClient) InitiateMultipartUpload(input *InitiateMultipartUploadInput) (output *InitiateMultipartUploadOutput, err error) {
-	if input == nil {
-		return nil, errors.New("InitiateMultipartUploadInput is nil")
-	}
-
-	if input.ContentType == "" && input.Key != "" {
-		if contentType, ok := mimeTypes[strings.ToLower(input.Key[strings.LastIndex(input.Key, ".")+1:])]; ok {
-			input.ContentType = contentType
-		}
-	}
-
-	output = &InitiateMultipartUploadOutput{}
-	err = obsClient.doActionWithBucketAndKey("InitiateMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output)
-	if err != nil {
-		output = nil
-	} else {
-		ParseInitiateMultipartUploadOutput(output)
-	}
-	return
-}
-
-func (obsClient ObsClient) UploadPart(_input *UploadPartInput) (output *UploadPartOutput, err error) {
-	if _input == nil {
-		return nil, errors.New("UploadPartInput is nil")
-	}
-
-	if _input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-
-	input := &UploadPartInput{}
-	input.Bucket = _input.Bucket
-	input.Key = _input.Key
-	input.PartNumber = _input.PartNumber
-	input.UploadId = _input.UploadId
-	input.ContentMD5 = _input.ContentMD5
-	input.SourceFile = _input.SourceFile
-	input.Offset = _input.Offset
-	input.PartSize = _input.PartSize
-	input.SseHeader = _input.SseHeader
-	input.Body = _input.Body
-
-	output = &UploadPartOutput{}
-	var repeatable bool
-	if input.Body != nil {
-		_, repeatable = input.Body.(*strings.Reader)
-		if _, ok := input.Body.(*readerWrapper); !ok && input.PartSize > 0 {
-			input.Body = &readerWrapper{reader: input.Body, totalCount: input.PartSize}
-		}
-	} else if sourceFile := strings.TrimSpace(input.SourceFile); sourceFile != "" {
-		fd, err := os.Open(sourceFile)
-		if err != nil {
-			return nil, err
-		}
-		defer fd.Close()
-
-		stat, err := fd.Stat()
-		if err != nil {
-			return nil, err
-		}
-		fileSize := stat.Size()
-		fileReaderWrapper := &fileReaderWrapper{filePath: sourceFile}
-		fileReaderWrapper.reader = fd
-
-		if input.Offset < 0 || input.Offset > fileSize {
-			input.Offset = 0
-		}
-
-		if input.PartSize <= 0 || input.PartSize > (fileSize-input.Offset) {
-			input.PartSize = fileSize - input.Offset
-		}
-		fileReaderWrapper.totalCount = input.PartSize
-		if _, err = fd.Seek(input.Offset, io.SeekStart); err != nil {
-			return nil, err
-		}
-		input.Body = fileReaderWrapper
-		repeatable = true
-	}
-	if repeatable {
-		err = obsClient.doActionWithBucketAndKey("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output)
-	} else {
-		err = obsClient.doActionWithBucketAndKeyUnRepeatable("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output)
-	}
-	if err != nil {
-		output = nil
-	} else {
-		ParseUploadPartOutput(output)
-		output.PartNumber = input.PartNumber
-	}
-	return
-}
-
-func (obsClient ObsClient) CompleteMultipartUpload(input *CompleteMultipartUploadInput) (output *CompleteMultipartUploadOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CompleteMultipartUploadInput is nil")
-	}
-
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-
-	var parts partSlice = input.Parts
-	sort.Sort(parts)
-
-	output = &CompleteMultipartUploadOutput{}
-	err = obsClient.doActionWithBucketAndKey("CompleteMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output)
-	if err != nil {
-		output = nil
-	} else {
-		ParseCompleteMultipartUploadOutput(output)
-	}
-	return
-}
-
-func (obsClient ObsClient) ListParts(input *ListPartsInput) (output *ListPartsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ListPartsInput is nil")
-	}
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-	output = &ListPartsOutput{}
-	err = obsClient.doActionWithBucketAndKey("ListParts", HTTP_GET, input.Bucket, input.Key, input, output)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
-func (obsClient ObsClient) CopyPart(input *CopyPartInput) (output *CopyPartOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CopyPartInput is nil")
-	}
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-	if strings.TrimSpace(input.CopySourceBucket) == "" {
-		return nil, errors.New("Source bucket is empty")
-	}
-	if strings.TrimSpace(input.CopySourceKey) == "" {
-		return nil, errors.New("Source key is empty")
-	}
-
-	output = &CopyPartOutput{}
-	err = obsClient.doActionWithBucketAndKey("CopyPart", HTTP_PUT, input.Bucket, input.Key, input, output)
-	if err != nil {
-		output = nil
-	} else {
-		ParseCopyPartOutput(output)
-		output.PartNumber = input.PartNumber
-	}
-	return
-}
