+++ openstack/obs/client_part.go
@@ -11,26 +11,20 @@ import (
 // ListMultipartUploads lists the multipart uploads.
 //
 // You can use this API to list the multipart uploads that are initialized but not combined or aborted in a specified bucket.
-func (obsClient ObsClient) ListMultipartUploads(input *ListMultipartUploadsInput, extensions ...extensionOptions) (output *ListMultipartUploadsOutput, err error) {
+func (obsClient ObsClient) ListMultipartUploads(input *ListMultipartUploadsInput) (output *ListMultipartUploadsOutput, err error) {
 	if input == nil {
 		return nil, errors.New("ListMultipartUploadsInput is nil")
 	}
 	output = &ListMultipartUploadsOutput{}
-	err = obsClient.doActionWithBucket("ListMultipartUploads", HTTP_GET, input.Bucket, input, output, extensions)
+	err = obsClient.doActionWithBucket("ListMultipartUploads", HTTP_GET, input.Bucket, input, output)
 	if err != nil {
 		output = nil
-	} else if output.EncodingType == "url" {
-		err = decodeListMultipartUploadsOutput(output)
-		if err != nil {
-			doLog(LEVEL_ERROR, "Failed to get ListMultipartUploadsOutput with error: %v.", err)
-			output = nil
-		}
 	}
 	return
 }
 
 // AbortMultipartUpload aborts a multipart upload in a specified bucket by using the multipart upload ID.
-func (obsClient ObsClient) AbortMultipartUpload(input *AbortMultipartUploadInput, extensions ...extensionOptions) (output *BaseModel, err error) {
+func (obsClient ObsClient) AbortMultipartUpload(input *AbortMultipartUploadInput) (output *BaseModel, err error) {
 	if input == nil {
 		return nil, errors.New("AbortMultipartUploadInput is nil")
 	}
@@ -38,7 +32,7 @@ func (obsClient ObsClient) AbortMultipartUpload(input *AbortMultipartUploadInput
 		return nil, errors.New("UploadId is empty")
 	}
 	output = &BaseModel{}
-	err = obsClient.doActionWithBucketAndKey("AbortMultipartUpload", HTTP_DELETE, input.Bucket, input.Key, input, output, extensions)
+	err = obsClient.doActionWithBucketAndKey("AbortMultipartUpload", HTTP_DELETE, input.Bucket, input.Key, input, output)
 	if err != nil {
 		output = nil
 	}
@@ -46,7 +40,7 @@ func (obsClient ObsClient) AbortMultipartUpload(input *AbortMultipartUploadInput
 }
 
 // InitiateMultipartUpload initializes a multipart upload.
-func (obsClient ObsClient) InitiateMultipartUpload(input *InitiateMultipartUploadInput, extensions ...extensionOptions) (output *InitiateMultipartUploadOutput, err error) {
+func (obsClient ObsClient) InitiateMultipartUpload(input *InitiateMultipartUploadInput) (output *InitiateMultipartUploadOutput, err error) {
 	if input == nil {
 		return nil, errors.New("InitiateMultipartUploadInput is nil")
 	}
@@ -58,18 +52,11 @@ func (obsClient ObsClient) InitiateMultipartUpload(input *InitiateMultipartUploa
 	}
 
 	output = &InitiateMultipartUploadOutput{}
-	err = obsClient.doActionWithBucketAndKey("InitiateMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
+	err = obsClient.doActionWithBucketAndKey("InitiateMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output)
 	if err != nil {
 		output = nil
 	} else {
 		ParseInitiateMultipartUploadOutput(output)
-		if output.EncodingType == "url" {
-			err = decodeInitiateMultipartUploadOutput(output)
-			if err != nil {
-				doLog(LEVEL_ERROR, "Failed to get InitiateMultipartUploadOutput with error: %v.", err)
-				output = nil
-			}
-		}
 	}
 	return
 }
@@ -79,7 +66,7 @@ func (obsClient ObsClient) InitiateMultipartUpload(input *InitiateMultipartUploa
 // After a multipart upload is initialized, you can use this API to upload a part to a specified bucket
 // by using the multipart upload ID. Except for the last uploaded part whose size ranges from 0 to 5 GB,
 // sizes of the other parts range from 100 KB to 5 GB. The upload part ID ranges from 1 to 10000.
-func (obsClient ObsClient) UploadPart(_input *UploadPartInput, extensions ...extensionOptions) (output *UploadPartOutput, err error) {
+func (obsClient ObsClient) UploadPart(_input *UploadPartInput) (output *UploadPartOutput, err error) {
 	if _input == nil {
 		return nil, errors.New("UploadPartInput is nil")
 	}
@@ -144,9 +131,9 @@ func (obsClient ObsClient) UploadPart(_input *UploadPartInput, extensions ...ext
 		repeatable = true
 	}
 	if repeatable {
-		err = obsClient.doActionWithBucketAndKey("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
+		err = obsClient.doActionWithBucketAndKey("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output)
 	} else {
-		err = obsClient.doActionWithBucketAndKeyUnRepeatable("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
+		err = obsClient.doActionWithBucketAndKeyUnRepeatable("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output)
 	}
 	if err != nil {
 		output = nil
@@ -158,7 +145,7 @@ func (obsClient ObsClient) UploadPart(_input *UploadPartInput, extensions ...ext
 }
 
 // CompleteMultipartUpload combines the uploaded parts in a specified bucket by using the multipart upload ID.
-func (obsClient ObsClient) CompleteMultipartUpload(input *CompleteMultipartUploadInput, extensions ...extensionOptions) (output *CompleteMultipartUploadOutput, err error) {
+func (obsClient ObsClient) CompleteMultipartUpload(input *CompleteMultipartUploadInput) (output *CompleteMultipartUploadOutput, err error) {
 	if input == nil {
 		return nil, errors.New("CompleteMultipartUploadInput is nil")
 	}
@@ -171,24 +158,17 @@ func (obsClient ObsClient) CompleteMultipartUpload(input *CompleteMultipartUploa
 	sort.Sort(parts)
 
 	output = &CompleteMultipartUploadOutput{}
-	err = obsClient.doActionWithBucketAndKey("CompleteMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
+	err = obsClient.doActionWithBucketAndKey("CompleteMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output)
 	if err != nil {
 		output = nil
 	} else {
 		ParseCompleteMultipartUploadOutput(output)
-		if output.EncodingType == "url" {
-			err = decodeCompleteMultipartUploadOutput(output)
-			if err != nil {
-				doLog(LEVEL_ERROR, "Failed to get CompleteMultipartUploadOutput with error: %v.", err)
-				output = nil
-			}
-		}
 	}
 	return
 }
 
 // ListParts lists the uploaded parts in a bucket by using the multipart upload ID.
-func (obsClient ObsClient) ListParts(input *ListPartsInput, extensions ...extensionOptions) (output *ListPartsOutput, err error) {
+func (obsClient ObsClient) ListParts(input *ListPartsInput) (output *ListPartsOutput, err error) {
 	if input == nil {
 		return nil, errors.New("ListPartsInput is nil")
 	}
@@ -196,15 +176,9 @@ func (obsClient ObsClient) ListParts(input *ListPartsInput, extensions ...extens
 		return nil, errors.New("UploadId is empty")
 	}
 	output = &ListPartsOutput{}
-	err = obsClient.doActionWithBucketAndKey("ListParts", HTTP_GET, input.Bucket, input.Key, input, output, extensions)
+	err = obsClient.doActionWithBucketAndKey("ListParts", HTTP_GET, input.Bucket, input.Key, input, output)
 	if err != nil {
 		output = nil
-	} else if output.EncodingType == "url" {
-		err = decodeListPartsOutput(output)
-		if err != nil {
-			doLog(LEVEL_ERROR, "Failed to get ListPartsOutput with error: %v.", err)
-			output = nil
-		}
 	}
 	return
 }
@@ -212,7 +186,7 @@ func (obsClient ObsClient) ListParts(input *ListPartsInput, extensions ...extens
 // CopyPart copy a part to a specified bucket by using a specified multipart upload ID.
 //
 // After a multipart upload is initialized, you can use this API to copy a part to a specified bucket by using the multipart upload ID.
-func (obsClient ObsClient) CopyPart(input *CopyPartInput, extensions ...extensionOptions) (output *CopyPartOutput, err error) {
+func (obsClient ObsClient) CopyPart(input *CopyPartInput) (output *CopyPartOutput, err error) {
 	if input == nil {
 		return nil, errors.New("CopyPartInput is nil")
 	}
@@ -227,7 +201,7 @@ func (obsClient ObsClient) CopyPart(input *CopyPartInput, extensions ...extensio
 	}
 
 	output = &CopyPartOutput{}
-	err = obsClient.doActionWithBucketAndKey("CopyPart", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
+	err = obsClient.doActionWithBucketAndKey("CopyPart", HTTP_PUT, input.Bucket, input.Key, input, output)
 	if err != nil {
 		output = nil
 	} else {
